services/search-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
opensearch-py[async]==2.4.2
aiohttp==3.9.5
pydantic-settings==2.3.4
httpx==0.27.0
pytest==9.0.3
pytest-asyncio==1.3.0
